- move AnswersConfigBuilder to it's own file - fix typos
